Refactor AddRecipe component; extract AddIngredients and AddInstructions into separate components, enhance styling and layout for better user experience
